separate conditionsParser from fieldParser into own file
