+++ commandParsers/fieldParser.js
@@ -1,9 +1,7 @@
-const _ = require('lodash')
 const {
     aggregateFunctionPattern,
     aggregateFunctionsNamePattern,
     arithmeticExpressionPattern,
-    comparisonOperatorPattern,
     modifiedArithmeticOperator,
     sortOrderKeywordPattern,
     distinctKeywordPattern,
@@ -11,78 +9,72 @@ const {
     stringFunctionsNamePattern,
     textInputPattern,
 } = require('../helpers/regex')
-const findIndexOfClosingBracket = require('./parserTools/findIndexOfClosingBracket')
 const {
     transformSelectInputArrayIntoFieldsArray,
-    transformSplitConditionsIntoConditionsArray,
     transformOrderByInputArrayIntoOrderByFieldsArray,
 } = require('./parserTools/arrayTransformationTools')
 const splitExpressionIntoArray = require('./parserTools/splitExpressionIntoArray')
 
 /**
- * Handles parsing of conditions from the given array.
- * @param {string[]} slicedCommandArray array containing the conditions
+ * Handles parsing of a field into field object.
+ * @param {String} field field as string
+ * @returns {object} field object
  */
-const parseConditions = (slicedCommandArray) => {
-    const conditionArray = transformSplitConditionsIntoConditionsArray(
-        slicedCommandArray
-    )
-
-    const conditions = { AND: [], OR: [] }
-
-    const indexOfAnd = _.indexOf(conditionArray, 'AND')
-    const indexOfOr = _.indexOf(conditionArray, 'OR')
-
-    let AndOrSwitch = indexOfOr < (indexOfAnd >= 0) ? 'AND' : 'OR'
-
-    for (let i = 0; i < conditionArray.length; i++) {
-        switch (conditionArray[i]) {
-            case '(': {
-                const indexOfClosingBracket = findIndexOfClosingBracket(
-                    conditionArray,
-                    i
-                )
-
-                conditions[AndOrSwitch] = conditions[AndOrSwitch].concat(
-                    parseConditions(
-                        conditionArray.slice(i + 1, indexOfClosingBracket)
-                    )
-                )
-                i = indexOfClosingBracket
-                break
+const parseField = (field) => {
+    switch (true) {
+        case stringFunctionPattern.test(field):
+            return {
+                type: 'stringFunction',
+                name: field.split('(')[0].toUpperCase(),
+                value: field,
+                param: parseParameterFromStringFunction(field),
             }
-            case ')':
-                break
-            case 'AND':
-                AndOrSwitch = 'AND'
-                break
-            case 'OR':
-                AndOrSwitch = 'OR'
-                break
-            default: {
-                if (comparisonOperatorPattern.test(conditionArray[i])) {
-                    const splitExpression = conditionArray[i].split(
-                        comparisonOperatorPattern
-                    )
-
-                    const condition = {
-                        left: parseField(splitExpression[0]),
-                        operator: splitExpression[1],
-                        right: parseField(splitExpression[2]),
-                    }
-                    conditions[AndOrSwitch] = conditions[AndOrSwitch].concat(
-                        condition
-                    )
-                } else {
-                    conditions[AndOrSwitch] = conditions[AndOrSwitch].concat(
-                        conditionArray[i]
-                    )
-                }
+        case aggregateFunctionPattern.test(field):
+            return {
+                type: 'aggregateFunction',
+                name: field.split('(')[0].toUpperCase(),
+                value: field,
+                param: parseParameterFromAggregateFunction(field),
             }
-        }
+        case /^\*$/.test(field):
+            return {
+                type: 'all',
+                value: field,
+            }
+        case arithmeticExpressionPattern.test(field):
+            return {
+                type: 'expression',
+                expressionParts: parseExpression(field),
+                value: field,
+            }
+        case textInputPattern.test(field):
+            return {
+                type: 'text',
+                value: field.replace(/'/g, ''),
+            }
+        case !isNaN(field):
+            return field
+                ? {
+                      type: 'integer',
+                      value: Number(field),
+                  }
+                : null
+        case modifiedArithmeticOperator.test(field):
+            return {
+                type: 'operator',
+                value: field === '**' ? '*' : field,
+            }
+        case sortOrderKeywordPattern.test(field):
+            return {
+                type: 'order',
+                value: field ? field.toLowerCase() : 'asc',
+            }
+        default:
+            return {
+                type: 'column',
+                value: field,
+            }
     }
-
-    return conditions
 }
 
 /**
@@ -151,68 +143,6 @@ const parseOrderByFields = (fieldArray) => {
 }
 
 /**
- * Handles parsing of a field into field object.
- * @param {String} field field as string
- * @returns {object} field object
- */
-const parseField = (field) => {
-    switch (true) {
-        case stringFunctionPattern.test(field):
-            return {
-                type: 'stringFunction',
-                name: field.split('(')[0].toUpperCase(),
-                value: field,
-                param: parseParameterFromStringFunction(field),
-            }
-        case aggregateFunctionPattern.test(field):
-            return {
-                type: 'aggregateFunction',
-                name: field.split('(')[0].toUpperCase(),
-                value: field,
-                param: parseParameterFromAggregateFunction(field),
-            }
-        case /^\*$/.test(field):
-            return {
-                type: 'all',
-                value: field,
-            }
-        case arithmeticExpressionPattern.test(field):
-            return {
-                type: 'expression',
-                expressionParts: parseExpression(field),
-                value: field,
-            }
-        case textInputPattern.test(field):
-            return {
-                type: 'text',
-                value: field.replace(/'/g, ''),
-            }
-        case !isNaN(field):
-            return field
-                ? {
-                      type: 'integer',
-                      value: Number(field),
-                  }
-                : null
-        case modifiedArithmeticOperator.test(field):
-            return {
-                type: 'operator',
-                value: field === '**' ? '*' : field,
-            }
-        case sortOrderKeywordPattern.test(field):
-            return {
-                type: 'order',
-                value: field ? field.toLowerCase() : 'asc',
-            }
-        default:
-            return {
-                type: 'column',
-                value: field,
-            }
-    }
-}
-
-/**
  * Handles parsing of a parameter from a string function.
  * @param {string[]} functionAsString array containing the string function
  */
@@ -239,7 +169,6 @@ const parseParameterFromAggregateFunction = (functionAsString) => {
 }
 
 module.exports = {
-    parseConditions,
     parseExpression,
     parseSelectFields,
     parseOrderByFields,
